touchup

+++ src/cli/formats.h
@@ -35,7 +35,7 @@ enum DecFormat {
 	DECFMT_WAV
 	//DECFMT_MKA_PCM
 };
-#define NUM_DECFMT		((size_t) 3u)
+#define NUM_DECFMT		((uint) 3u)
 #define DECFMT_NAME_ARRAY	{"raw", "w64", "wav"}
 
 enum IntType {
